+++ scripts/macros/spells/deathWard.js
@@ -1,5 +1,4 @@
 import {effectUtils, genericUtils, itemUtils} from '../../utils.js';
-
 async function use({workflow}) {
     let effectData = {
         name: workflow.item.name,